Syntax optimization + wikipedia unblocking
Optimized some of the syntax for the first entries in the list based on what I learned on the Pinterest uBlacklist blocklist. https://github.com/rjaus/ublacklist-pinterest/blob/main/ublacklist-pinterest-ext.txt

The appstore stuff, for example, only needs the app ID: it's a constant and if you remove it, you can't access the page. I don't know all of the sites' structures, especially the German ones, so you might have to correct my changes. Your lists are long, I'm quite lazy, and you know them better than I do, so I hope this can help you maintaining them!

I also strongly believe that Wikipedia entries shouldn't be on any blocklists: it's not owned by AS and is where I learned about all their terrible politics and actions. Currently, there are political parties and companies that are lobbying for Wikipedia censorship *because* they publish these informations! It's a counter-power, not a propaganda outlet (when the mods are awake).